+++ package/src/components/AudioPlayer/Context/UIContext.ts
@@ -7,12 +7,15 @@ import {
   VolumeSliderPlacement,
 } from "./StateContext";
 
+export type ColorScheme = "light" | "dark";
+
 export interface UIContext {
   activeUI: ActiveUI;
   playListPlacement: PlayListPlacement;
   playerPlacement?: PlayerPlacement;
   interfacePlacement?: InterfacePlacement;
   volumeSliderPlacement?: VolumeSliderPlacement;
+  colorScheme?: ColorScheme;
 }
 
 export const uiContext = createContext<UIContext | null>(null);